feat: Add path mapping for vite in tsconfig.json
This commit adds a new "paths" configuration in the tsconfig.json file to enable path mapping for the "vite" module. By specifying the path to the "vite" module in the "paths" configuration, the TypeScript compiler will be able to resolve the module correctly during the build process. This improves the overall build performance and ensures that the application can use the latest version of the "vite" module.