feat: Add Twitter share button to layout
This commit adds the Twitter share button to the layout component in order to allow users to easily share the page on Twitter. The share button is implemented using the ShareButton component and is displayed in both the main content area and the mobile view. This enhancement improves the social sharing capabilities of the website and encourages users to share the content with their followers.

Note: This commit message follows the established convention of starting with a verb in the present tense and providing a clear and concise description of the changes made.